черновик. DirectoryDownloader теперь рекурсивно скачивает всю папку. исправлена ошибка в классе CaptionView.

+++ PhotoBrowser/LoadingDelegate.h
@@ -15,6 +15,6 @@
 - (void)handleErrorNotification:(id)sender;
 
 @optional
-- (void)handleLoadingProgressNotification:(NSUInteger)value;
+- (void)handleLoadingProgressNotification:(double)value;
 
 @end
